refactoring(EDCO-45): EIcon for Feature
