+++ src/stories/Feature/featureCommons.tsx
@@ -1,20 +1,19 @@
-import AccountBoxIcon from '@mui/icons-material/AccountBox';
-import { type FeatureItem } from 'components/Feature/Feature';
+import { type FeatureItem } from 'components/Feature/FeatureStackItem';
 
 export const itemMockWithoutLink: FeatureItem[] = [
   {
-    icon: <AccountBoxIcon />,
+    stackIcon: { icon: 'AccountBox' },
     title: 'Sicuro',
     subtitle:
       'I tuoi dati e le informazioni contenute negli atti sono al sicuro',
   },
   {
-    icon: <AccountBoxIcon />,
+    stackIcon: { icon: 'AccountBox' },
     title: 'Conveniente',
     subtitle: 'Un risparmio di denaro per tutti: cittadini imprese ed enti',
   },
   {
-    icon: <AccountBoxIcon />,
+    stackIcon: { icon: 'AccountBox' },
     title: 'Immediato',
     subtitle:
       "I tempi di attesa tra l'invio e la ricezione delle notifiche spariscono",
